requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.1.0
pydantic==2.6.4
httpx==0.27.0
pytest==8.3.3